Added test confirming asymmetry of the operator c
